+++ src/components/ProfileDropdown/profileDropdown.module.css
@@ -1,17 +1,3 @@
-.pageContainer {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  padding: 1rem 1.5rem 0 calc(300px + 2rem + 1.5rem);
-}
-
-.expand {
-  padding-left: 4rem;
-  animation: moveLeft 0.5s ease-in-out;
-}
-.avatarStyle {
-  border-radius: 100%;
-}
 .profileContainer {
   border: none;
   padding: 2.1rem 0.5rem;
@@ -57,45 +43,6 @@
   display: block;
   text-transform: capitalize;
 }
-
-.contract {
-  padding-left: calc(300px + 2rem + 1.5rem);
-  animation: moveRight 0.5s ease-in-out;
-}
-
-.collapseSidebarButton {
-  position: fixed;
-  height: 40px;
-  bottom: 0;
-  z-index: 9999;
-  width: calc(300px + 2rem);
-  background-color: rgba(245, 245, 245, 0.7);
-  color: black;
-  border: none;
-  border-radius: 0px;
-}
-
-.collapseSidebarButton:hover,
-.opendrawer:hover {
-  opacity: 1;
-  color: black !important;
-}
-.opendrawer {
-  position: fixed;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  top: 0;
-  left: 0;
-  width: 40px;
-  height: 100vh;
-  z-index: 9999;
-  background-color: rgba(245, 245, 245);
-  border: none;
-  border-radius: 0px;
-  margin-right: 20px;
-  color: black;
-}
 .profileDropdown {
   background-color: transparent !important;
 }
@@ -114,34 +61,6 @@
   border-top: none !important;
   border-bottom: none !important;
 }
-
-@media (max-width: 1120px) {
-  .contract {
-    padding-left: calc(250px + 2rem + 1.5rem);
-  }
-  .collapseSidebarButton {
-    width: calc(250px + 2rem);
-  }
-}
-
-/* For tablets */
-@media (max-width: 820px) {
-  .pageContainer {
-    padding-left: 2.5rem;
-  }
-
-  .opendrawer {
-    width: 25px;
-  }
-
-  .contract,
-  .expand {
-    animation: none;
-  }
-
-  .collapseSidebarButton {
-    width: 100%;
-    left: 0;
-    right: 0;
-  }
+.avatarStyle {
+  border-radius: 100%;
 }
